logical: extract error helpers to sqlwriter package
Move the canDlqError function from the logical package to the sqlwriter
package as an exported CanDLQError function and add an IsLwwLoser helper
that detects ConditionFailedErrors with OriginTimestampOlderThan set.
This makes error classification logic reusable by other LDR writers.

Release note: None
Epic: CRDB-57649
