+++ pkg/crosscluster/logical/logical_replication_writer_processor.go
@@ -9,14 +9,13 @@ import (
 	"context"
 	"fmt"
 	"hash/fnv"
-	"regexp"
 	"slices"
-	"strings"
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/cdcevent"
 	"github.com/cockroachdb/cockroach/pkg/clusterversion"
 	"github.com/cockroachdb/cockroach/pkg/crosscluster"
+	"github.com/cockroachdb/cockroach/pkg/crosscluster/logical/sqlwriter"
 	"github.com/cockroachdb/cockroach/pkg/crosscluster/replicationutils"
 	"github.com/cockroachdb/cockroach/pkg/crosscluster/streamclient"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
@@ -35,7 +34,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/isql"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/rowexec"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
@@ -1121,7 +1119,7 @@ func (lrw *logicalReplicationWriterProcessor) maybeDLQ(
 	applyErr error,
 	eligibility retryEligibility,
 ) error {
-	if err := canDlqError(applyErr); err != nil {
+	if err := sqlwriter.CanDLQError(applyErr); err != nil {
 		return errors.Wrapf(err, "dlq eligibility %s", eligibility)
 	}
 	if log.V(1) || logAllDLQs {
@@ -1144,42 +1142,6 @@ func (lrw *logicalReplicationWriterProcessor) maybeDLQ(
 		lrw.metrics.DLQedDueToErrType.Inc(1)
 	}
 	return lrw.dlqClient.Log(ctx, lrw.spec.JobID, event, row, applyErr, eligibility)
-}
-
-var internalPgErrors = func() *regexp.Regexp {
-	codePrefixes := []string{
-		// Section: Class 57 - Operator Intervention
-		"57",
-		// Section: Class 58 - System Error
-		"58",
-		// Section: Class 25 - Invalid Transaction State
-		"25",
-		// Section: Class 2D - Invalid Transaction Termination
-		"2D",
-		// Section: Class 40 - Transaction Rollback
-		"40",
-		// Section: Class XX - Internal Error
-		"XX",
-		// Section: Class 58C - System errors related to CockroachDB node problems.
-		"58C",
-	}
-	return regexp.MustCompile(fmt.Sprintf("^(%s)", strings.Join(codePrefixes, "|")))
-}()
-
-// canDlqError returns true if the error should send a row to the DLQ. We don't
-// want to DLQ rows that failed to apply because of some internal problem like
-// an unavailable range. The idea is it is better to fail the processor so the
-// job backs off until the system is healthy.
-func canDlqError(err error) error {
-	// If the error is not from the SQL layer, then we don't want to DQL it.
-	if !pgerror.HasCandidateCode(err) {
-		return errors.Wrap(err, "can only DLQ errors with pg codes")
-	}
-	code := pgerror.GetPGCode(err)
-	if internalPgErrors.MatchString(code.String()) {
-		return errors.Wrap(err, "unable to DLQ pgcode that indicates an internal or retryable error")
-	}
-	return nil
 }
 
 type batchStats struct {
